+++ NEWS.md
@@ -3,7 +3,8 @@
 * added R (>= 4.1.0) dependency for use of base pipe operator
 * added CRAN DOI badge
 * added 2021 and 2022 typhoon data
-* added utility unexported function to download cyclone reports
+* added CITATION.cff
+* added an unexported utility function to download cyclone reports
 
 
 # bagyo 0.1.1