+++ NEWS.md
@@ -1,8 +1,10 @@
-# bagyo v0.1.1.9000
+# bagyo v0.2.0
 
+* added R (>= 4.1.0) dependency for use of base pipe operator
 * added CRAN DOI badge
-* added 2021 typhoon data
-* added utility unexported function to download cyclone reports
+* added 2021 and 2022 typhoon data
+* added CITATION.cff
+* added an unexported utility function to download cyclone reports
 
 
 # bagyo 0.1.1